fix PlotStartTime
