+++ src/plot.ts
@@ -1,7 +1,5 @@
 import { PlotLineMapControl } from "./plot_line_map_control";
 
-// (() => {
-// declare let THREE: never;
 
 function calculateNumberOfDigits(interval) {
   let num = Math.floor(Math.log(interval) / Math.log(10));
@@ -57,172 +55,9 @@ var plot_animate = [];
 //   }
 // }
 
-var array = -1;
-
-
-
-
 
 var animation_line = [];
-var current_line_vec_animation = [];
-var s_geometry;
-var s_material;
-var sphere;
-var phase_index;
-var phase;
-var vec;
 
-function add_plot_each(phase_index_array, axes, line: PlotLine, width, color, dt, parameter_condition_list, current_param_idx, current_line_vec) {
-  try {
-    while (true) {
-      if (line.plot_ready) {
-        line.plotting = false;
-        console.log("Plot is interrupted");
-        PlotStartTime = undefined;
-        return;
-      }
-
-      // phase_index_array is used to implement dfs without function call.
-      phase_index = phase_index_array[phase_index_array.length - 1];
-      phase = phase_index.phase;
-      vec = phase_to_line_vectors(phase, parameter_condition_list[current_param_idx], axes, dt);
-      current_line_vec = current_line_vec.concat(vec);
-      let vec_animation = phase_to_line_vectors(phase, parameter_condition_list[current_param_idx], axes, 0.01);
-      current_line_vec_animation = current_line_vec_animation.concat(vec_animation);
-      if (phase.children.length == 0) {
-        array += 1;
-        // on leaves
-
-        var cylindersGeometry = new THREE.Geometry();
-        let scaledWidth = 0.5 * width / graph.camera.zoom;
-        var addCylinder = function (startPos: THREE.Vector3, endPos: THREE.Vector3) {
-          let directionVec = endPos.clone().sub(startPos);
-          const height = directionVec.length();
-          directionVec.normalize();
-          let cylinderMesh = new THREE.Mesh(new THREE.CylinderGeometry(scaledWidth, scaledWidth, height + scaledWidth, 8, 1));
-
-          const upVec = new THREE.Vector3(0, 1, 0);
-          const rotationAxis = upVec.clone().cross(directionVec).normalize();
-          const rotationAngle = Math.acos(upVec.dot(directionVec));
-
-          const newpos = startPos.clone().lerp(endPos, 0.5);
-          cylinderMesh.position.set(newpos.x, newpos.y, newpos.z);
-          cylinderMesh.setRotationFromAxisAngle(rotationAxis, rotationAngle);
-
-          cylinderMesh.updateMatrix();
-          const geometry = cylinderMesh.geometry instanceof THREE.Geometry ? cylinderMesh.geometry : new THREE.Geometry().fromBufferGeometry(cylinderMesh.geometry);
-          cylindersGeometry.merge(geometry, cylinderMesh.matrix);
-        };
-
-        const dottedLength = 10.0 / graph.camera.zoom;
-        for (var i = 0; i + 1 < current_line_vec.length; i++) {
-          if ('isPP' in current_line_vec[i + 1]) {
-            const posBegin = current_line_vec[i];
-            const posEnd = current_line_vec[i + 1];
-            let directionVec = posEnd.clone().sub(posBegin);
-            const lineLength = directionVec.length();
-            directionVec.normalize();
-            const numOfDots = lineLength / dottedLength;
-            for (var j = 1; j + 1 < numOfDots; j += 2) {
-              addCylinder(
-                posBegin.clone().add(directionVec.clone().multiplyScalar(j * dottedLength)),
-                posBegin.clone().add(directionVec.clone().multiplyScalar((j + 1) * dottedLength))
-              );
-            }
-          }
-          else if (!current_line_vec[i].equals(current_line_vec[i + 1])) {
-            addCylinder(current_line_vec[i], current_line_vec[i + 1]);
-          }
-        }
-
-        let three_line = new THREE.Mesh(
-          cylindersGeometry,
-          new THREE.MeshBasicMaterial({ color: color[current_param_idx] })
-        );
-        three_line.isLine = true;
-        graph.scene.add(three_line);
-        line.plot.push(three_line);
-
-        animation_line[array] = (current_line_vec_animation);
-        animation_line[array].color = (color[current_param_idx]);
-        if (animation_line.maxlen < current_line_vec_animation.length) {
-          animation_line.maxlen = current_line_vec_animation.length;
-        }
-        s_geometry = new THREE.SphereGeometry(0.1);
-        s_material = new THREE.MeshBasicMaterial({ color: color[current_param_idx] });
-        sphere = new THREE.Mesh(s_geometry, s_material);
-        sphere.position.set(0, 0, 0);
-        graph.scene.add(sphere);
-        plot_animate[array] = (sphere);
-        current_line_vec = [];
-        current_line_vec_animation = [];
-        phase_index_array.pop();
-        phase_index = phase_index_array[phase_index_array.length - 1];
-        ++(phase_index.index);
-        phase = phase_index.phase;
-        if (animation_line.length > 4) {
-          console.log("a");
-        }
-      }
-      while (true) {
-        var to_next_child = false;
-        // search next child to plot
-        for (; phase_index.index < phase.children.length; phase_index.index++) {
-          var child = phase.children[phase_index.index];
-          var included_by_parameter_condition = check_parameter_condition(child.parameter_maps, parameter_condition_list[current_param_idx]);
-          if (included_by_parameter_condition) {
-            phase_index_array.push({ phase: child, index: 0 });
-            var current_time = new Date().getTime();
-            if (current_time - line.last_plot_time >= 200) {
-              line.last_plot_time = current_time;
-              // use setTimeout to check event queue
-              requestAnimationFrame(function () {
-                add_plot_each(phase_index_array, axes, line, width, color, dt, parameter_condition_list, current_param_idx, current_line_vec)
-              });
-              return;
-            }
-            else to_next_child = true;
-            break;
-          }
-        }
-        if (to_next_child) break;
-
-
-        // Plot for this current_param_idx is completed.
-        if (phase_index_array.length == 1) {
-          ++current_param_idx;
-          if (current_param_idx >= parameter_condition_list.length) {
-            // Plot is completed.
-            line.plotting = false;
-            checkAndStopPreloader();
-            return;
-          }
-          else {
-            // setTimeout(function()
-            //             {add_plot_each([{phase:phase_index_array[0].phase, index:0}], axes, line, width, color, dt, parameter_condition_list, current_param_idx, [])
-            //             }, 0);
-            phase_index_array[0].index = 0;
-            break;
-          }
-        } else {
-          // go to parent phase
-          phase_index_array.pop();
-          phase_index = phase_index_array[phase_index_array.length - 1];
-          ++(phase_index.index);
-          phase = phase_index.phase;
-        }
-      }
-    }
-  }
-  catch (ex) {
-    console.log(ex);
-    console.log(ex.stack);
-    showToast("Plot failed: " + ex.name +
-      "(" + ex.message + ")", 3000, "red darken-4");
-    line.plotting = false;
-    checkAndStopPreloader();
-  }
-}
 
 
 function divideParameter(parameter_map) {
@@ -374,17 +209,6 @@ function check_parameter_condition(parameter_maps, parameter_condition_list) {
   return false;
 }
 
-function checkAndStopPreloader() {
-  var table = document.getElementById("graph_axis_table");
-  if (!PlotLineMapControl.isAllReady()) return;
-  var current_time = new Date().getTime();
-  if (PlotStartTime == undefined || current_time - PlotStartTime >= 1000) {
-    showToast("Plot finished.", 1000, "blue");
-  }
-  PlotStartTime = undefined;
-  graph.renderer.render(graph.scene, graph.camera);
-  stopPreloader();
-}
 
 function remove_plot(line) {
   if (line.plot != undefined) {
